Normative: Access properties in alphabetical order
This aligns ToDurationRecord with ToTemporalDurationRecord, which accesses
properties on the passed-in object in alphabetical order. It makes the
order no longer rely on the order of rows in the Duration components
table.

This takes the approach of "unrolling" the loop. Another approach would be
to sort the list of property names, but the ECMA-262 editors have said
they prefer this approach.

Closes: #95