Editorial: Simplify tracking of separated parts
Only hours, minutes, seconds, and sub-seconds can have a "numeric" or
"2-digit" style. And if a unit is numeric-like, then all smaller units
are also numeric-like. And numeric-like seconds or sub-seconds are
always the last element, any smaller units are part of the fractional
substring.

That means we can reorder the steps as follows:
- The `separated` variable is set to `true` when the first numeric-like
  unit is encountered.
- Any following numeric-like unit prepends the `separator` string when
  `separated` is `true`.