Editorial: Added "fractional" style that's only used internally.
"numeric" behaves two very different ways: if it's used on hours,
minutes, or seconds it formats the values with separator characters
(i.e. : in 'en') between them. If it's used on milliseconds,
microseconds, or nanoseconds, it formats the values as a fractional
component of the smallest non-"numeric" or "2-digit" unit.

By treating the "numeric" value as "fractional" in all cases except when
creating the DurationFormat object and in .resolvedOptions(), simplifies
the code somewhat. It also makes it relatively easier to actually
introduce a new observable "fractional" style.