Editorial: updated resolvedOptions and Table 3 to use new "fractional" style for subsecond units.
This is introduced to simplify the logic for handling "numeric" units in
GetDurationUnitOptions and PartitionDurationFormatPattern, as this style means
something different for subsecond units than it does for hours, minutes, and
seconds.

This change is non-observable: users still use "numeric" to specify
the style in all cases, and "fractional" is converted back to "numeric" in
resolvedOptions.